Update swoop.inc.c (#767)
"Stun" was misleading. The swooper touches a wall, then it will slowly change direction so it's not stuck on that wall. When it does so, it could take a second to fully turn around and get off the wall, so that countdown is there so the swooper doesn't start turning back towards the wall again.

ALSO! We're supposed to change all the hex to decimal, right? At least it makes sense here cuz of all the zeroes lying around after converting them.